# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Self-Hosted Video Chat API
  version: 0.1.0
  description: |
    Core REST API contracts for the self-hosted video chat platform.
    This MVP specification covers authentication, user discovery, friend management,
    room metadata, messaging summaries, WebRTC negotiation helpers, and health checks.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.example.com
    description: Production placeholder

tags:
  - name: auth
    description: Authentication and session lifecycle endpoints.
  - name: users
    description: User discovery and profile queries.
  - name: friends
    description: Friend requests, approvals, and roster listing.
  - name: rooms
    description: Spatial audio/video rooms available to users.
  - name: messages
    description: Chat message retrieval and delivery helpers.
  - name: webrtc
    description: Signaling helpers for establishing peer connections.
  - name: turn
    description: TURN/STUN credential distribution.
  - name: health
    description: Operational readiness and monitoring probes.

paths:
  /auth/register:
    post:
      summary: Register a new account
      description: Create a new user account.
      operationId: register
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegister'
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '400':
          description: Invalid payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email or username already taken.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Authenticate with username or email
      description: Exchange credentials for a session cookie (web) or JWT pair (desktop clients).
      operationId: login
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLogin'
      responses:
        '200':
          description: Credentials accepted; session established.
          headers:
            Set-Cookie:
              description: Present for web logins; issues httpOnly session cookie with CSRF token in body.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is blocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      summary: Rotate refresh token
      description: Exchange a valid refresh token for a new access token and refresh token.
      operationId: refreshAuth
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefresh'
      responses:
        '200':
          description: New token pair issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRefreshResponse'
        '401':
          description: Refresh token invalid or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is blocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Terminate the current session
      description: Invalidates the active session cookie or refresh token for the caller's device.
      operationId: logout
      tags:
        - auth
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLogout'
      responses:
        '204':
          description: Session terminated.
        '401':
          description: Session not found or already invalidated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: CSRF or refresh token validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      summary: Retrieve the current authenticated user
      description: Returns the authenticated user if the requester holds a valid session.
      operationId: getCurrentUser
      tags:
        - auth
      responses:
        '200':
          description: Authenticated user payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: No valid session available.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/search:
    get:
      summary: Search for users
      description: Case-insensitive search across usernames and display names.
      operationId: searchUsers
      tags:
        - users
      parameters:
        - in: query
          name: q
          description: Free-form search query (minimum 2 characters).
          required: true
          schema:
            type: string
            minLength: 2
        - in: query
          name: limit
          description: Maximum number of results to return.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Matching users returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '400':
          description: Query too short or parameters invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /friends/request:
    post:
      summary: Send a friend request
      description: Initiates a new friend request to another user.
      operationId: requestFriend
      tags:
        - friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequestPayload'
      responses:
        '202':
          description: Friend request queued for review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
        '400':
          description: Request invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request already exists or users already connected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /friends/accept:
    post:
      summary: Accept a pending friend request
      description: Confirms a friend relationship and moves it into an accepted state.
      operationId: acceptFriend
      tags:
        - friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendDecisionPayload'
      responses:
        '200':
          description: Friend request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /friends/decline:
    post:
      summary: Decline a pending friend request
      description: Rejects a friend request and leaves both users unconnected.
      operationId: declineFriend
      tags:
        - friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendDecisionPayload'
      responses:
        '200':
          description: Friend request declined.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /friends:
    get:
      summary: List friend relationships
      description: Returns both accepted friends and pending requests for the current user.
      operationId: listFriends
      tags:
        - friends
      parameters:
        - in: query
          name: status
          description: Filter by relationship status.
          required: false
          schema:
            type: string
            enum:
              - pending
              - accepted
              - blocked
      responses:
        '200':
          description: Friends returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Friend'
        '401':
          description: Session invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /rooms:
    get:
      summary: List rooms available to the current user
      description: Provides metadata about joinable rooms for lobby selection.
      operationId: listRooms
      tags:
        - rooms
      responses:
        '200':
          description: Rooms returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
        '401':
          description: Session invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /rooms/{roomId}/messages:
    get:
      summary: Fetch recent messages for a room
      description: Returns paginated chat messages for the specified room.
      operationId: listRoomMessages
      tags:
        - messages
      parameters:
        - in: path
          name: roomId
          required: true
          description: Unique identifier of the room.
          schema:
            type: string
            format: uuid
        - in: query
          name: before
          required: false
          description: Return messages sent before this ISO timestamp.
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          required: false
          description: Maximum number of messages to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Messages returned for the requested room.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - hasMore
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  hasMore:
                    type: boolean
                    description: Indicates if more messages are available when paginating backwards.
        '401':
          description: Session invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Room not found or access denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webrtc/ice-config:
    get:
      summary: Retrieve ICE server configuration
      description: Provides STUN/TURN credentials for establishing peer connections.
      operationId: getIceConfig
      tags:
        - webrtc
        - turn
      responses:
        '200':
          description: ICE configuration payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IceConfig'
        '401':
          description: Session invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /healthz:
    get:
      summary: Health probe
      description: Returns service readiness information for orchestrators.
      operationId: healthCheck
      tags:
        - health
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

components:
  schemas:
    User:
      type: object
      required:
        - id
        - username
        - displayName
        - avatarUrl
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 3f2a6e82-4f14-4dcd-ab1a-2c8c8e9b444c
        username:
          type: string
          minLength: 3
          maxLength: 32
          example: maxim
        displayName:
          type: string
          example: Maxim Runkov
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.com/avatars/maxim.png
        bio:
          type: string
          nullable: true
          example: Loves self-hosted infrastructure.
        createdAt:
          type: string
          format: date-time
          example: 2024-01-18T14:12:03Z

    AuthUser:
      type: object
      required:
        - id
        - username
        - email
        - is_blocked
      properties:
        id:
          type: integer
          example: 42
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: maxim
        email:
          type: string
          format: email
          example: maxim@example.com
        is_blocked:
          type: boolean
          example: false

    AuthRegister:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
          example: maxim@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: maxim
        password:
          type: string
          format: password
          minLength: 8
          example: use-a-strong-password

    AuthDevice:
      type: object
      description: Optional device metadata used to scope refresh tokens.
      properties:
        identifier:
          type: string
          maxLength: 128
          nullable: true
          example: 58c5d1c2f2b64d589a3b85ab
        kind:
          type: string
          enum:
            - web
            - desktop
            - tauri
          default: web
        display_name:
          type: string
          maxLength: 100
          nullable: true
          example: MacBook Pro

    AuthLogin:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Username or email address.
          example: maxim
        password:
          type: string
          format: password
          minLength: 8
          example: use-a-strong-password
        device:
          $ref: '#/components/schemas/AuthDevice'

    AuthLoginResponse:
      type: object
      properties:
        csrf_token:
          type: string
          nullable: true
          example: _pMtJY45yFomN6c8tZp0aQY8w4fT7b0z58aW2g
        session_expires_in:
          type: integer
          nullable: true
          example: 604800
        access_token:
          type: string
          nullable: true
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          nullable: true
          example: K3Q4YjsU1q0zvdZ-bm1FvQ
        expires_in:
          type: integer
          nullable: true
          example: 900
        refresh_expires_in:
          type: integer
          nullable: true
          example: 2592000
        device_id:
          type: string
          nullable: true
          example: 58c5d1c2f2b64d589a3b85ab
        token_type:
          type: string
          enum:
            - bearer
          nullable: true

    AuthRefresh:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          minLength: 16
          example: K3Q4YjsU1q0zvdZ-bm1FvQ
        device_id:
          type: string
          maxLength: 128
          nullable: true
          example: 58c5d1c2f2b64d589a3b85ab

    AuthRefreshResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_in
        - refresh_expires_in
        - device_id
        - token_type
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: M2J6sAFVY6hFIZq6Q1l5sg
        expires_in:
          type: integer
          example: 900
        refresh_expires_in:
          type: integer
          example: 2592000
        device_id:
          type: string
          example: 58c5d1c2f2b64d589a3b85ab
        token_type:
          type: string
          enum:
            - bearer

    AuthLogout:
      type: object
      properties:
        refresh_token:
          type: string
          minLength: 16
          nullable: true
          example: K3Q4YjsU1q0zvdZ-bm1FvQ
        device_id:
          type: string
          maxLength: 128
          nullable: true
          example: 58c5d1c2f2b64d589a3b85ab
    Friend:
      type: object
      required:
        - id
        - requesterId
        - addresseeId
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        requesterId:
          type: string
          format: uuid
        addresseeId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending
            - accepted
            - declined
            - blocked
        createdAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
          nullable: true
        requester:
          $ref: '#/components/schemas/User'
        addressee:
          $ref: '#/components/schemas/User'

    Room:
      type: object
      required:
        - id
        - name
        - ownerId
        - participantCount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Daily Standup
        description:
          type: string
          nullable: true
        ownerId:
          type: string
          format: uuid
        isPrivate:
          type: boolean
          default: false
        participantCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true

    Message:
      type: object
      required:
        - id
        - roomId
        - senderId
        - body
        - sentAt
      properties:
        id:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        body:
          type: string
          example: Ready for the call?
        sentAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
          nullable: true
        attachments:
          type: array
          items:
            type: string
            format: uri

    IceConfig:
      type: object
      required:
        - ttlSeconds
        - servers
      properties:
        ttlSeconds:
          type: integer
          minimum: 30
          example: 3600
        servers:
          type: array
          items:
            type: object
            required:
              - urls
            properties:
              urls:
                oneOf:
                  - type: string
                  - type: array
                    items:
                      type: string
              username:
                type: string
                nullable: true
              credential:
                type: string
                nullable: true

    FriendRequestPayload:
      type: object
      required:
        - targetUserId
      properties:
        targetUserId:
          type: string
          format: uuid

    FriendDecisionPayload:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
          format: uuid

    HealthStatus:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          example: ok
        timestamp:
          type: string
          format: date-time
        detail:
          type: string
          nullable: true

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
          example: One or more fields failed validation.
        fields:
          type: object
          additionalProperties:
            type: string
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session